--- Components/CartIcon.tsx
@@ -1,20 +1,0 @@
-// import React from 'react';
-// import { TouchableOpacity, Text } from 'react-native';
-// import { useCart } from '../src/Context/CartContext';
-
-// const CartIcon: React.FC<{ navigation: any }> = ({ navigation }) => {
-//   const { cartCount } = useCart();
-
-//   const handleOpenCart = () => {
-//     navigation.navigate('CartScreen');
-//   };
-
-//   return (
-//     <TouchableOpacity onPress={handleOpenCart}>
-//       <Text>Cart</Text>
-//       <Text>{cartCount}</Text>
-//     </TouchableOpacity>
-//   );
-// };
-
-// export default CartIcon;
