--- Components/CartScreen.tsx
@@ -1,21 +1,0 @@
-// import React from 'react';
-// import { View, Text } from 'react-native';
-// import { useCart } from '../src/Context/CartContext';
-
-// const CartScreen: React.FC = () => {
-//   const { cartItems, cartTotal } = useCart();
-
-//   return (
-//     <View>
-//       <Text>Cart Items:</Text>
-//       {cartItems.map((item:any, index:any) => (
-//         <Text key={index}>
-//           {item.name} - {item.price}
-//         </Text>
-//       ))}
-//       <Text>Total: {cartTotal}</Text>
-//     </View>
-//   );
-// };
-
-// export default CartScreen;
